+++ src/main/java/com/jamersc/springboot/financialhub/model/Sample.java
@@ -11,5 +11,5 @@ public class Sample {
 
     private String firstName;
     private String lastName;
-    private int age;
+    private Long age;
 }
